requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
distro==1.9.0
filelock==3.20.3
Flask==3.1.2
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
fonttools==4.59.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.78.0
grpcio-status==1.71.2
gunicorn==25.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.47.1
networkx==3.6.1
nltk==3.9.2
numba==0.63.1
numpy==2.4.2
openai==2.20.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==6.5.2
proto-plus==1.27.1
protobuf==5.29.6
pyarrow==21.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.4
rich==14.3.2
rpds-py==0.26.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
setuptools==82.0.0
shap==0.50.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
streamlit==1.47.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.1
tqdm==4.67.3
transformers==5.1.0
typer==0.23.0
typer-slim==0.23.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
